Since the branches had diverged so much, it was simpler to apply a heavy hammer
of "git merge -s ort -X theirs" and then deal with the fallout.

+++ sys/sys/ktrace.h
@@ -68,7 +68,7 @@ struct ktr_header {
 	pid_t	ktr_pid;		/* process id */
 	char	ktr_comm[MAXCOMLEN + 1];/* command name */
 	struct	timespec ktr_time;	/* timestamp */
-	intptr_t	ktr_tid;	/* thread id */
+	long	ktr_tid;		/* thread id */
 	int	ktr_cpu;		/* cpu id */
 };
